backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai==1.100.2
pillow==11.3.0
python-dotenv==1.1.1
